# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Email Automation Simulator API",
    "version": "1.0.0",
    "description": "API for simulating email warmups, interactions, and analytics."
  },
  "paths": {
    "/api/accounts": {
      "get": {
        "summary": "Fetch Account Details",
        "responses": { "200": { "description": "A list of accounts" } }
      },
      "post": {
        "summary": "Create Account",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string" },
                  "name": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Account created" } }
      }
    },
...
  }
}